Add esbuild-based build and support HACS
Add esbuild-based build process and HACS support.
